fix: disable refreshing of extensions when runtime management is disabled (#10204)

+++ src/Core/Framework/Store/Api/ExtensionStoreActionsController.php
@@ -42,6 +42,10 @@ class ExtensionStoreActionsController extends AbstractController
     #[Route(path: '/api/_action/extension/refresh', name: 'api.extension.refresh', methods: ['POST'])]
     public function refreshExtensions(Context $context): Response
     {
+        if (!$this->runtimeExtensionManagementAllowed) {
+            return new Response('', Response::HTTP_NO_CONTENT);
+        }
+
         $this->pluginService->refreshPlugins($context, new NullIO());
 
         $this->cache->delete(StoreClient::EXTENSION_LIST_CACHE);
